Online Book Reselling allows users to buy and sell used books. Made using HTML, CSS and C#
Online Book Reselling System will provide the buyers and sellers the online shopping facility to buy and even upload their respective books on our website. As this is a website based application the user has to and register themselves on our website .

Basically, this is a virtual bookstore where books can be bought from the comfort of home through the internet.